64 grade the quizzes

+++ app/src/main/java/com/example/learnit/ui/feature/courses/quiz/model/QuizResultModel.kt
@@ -3,5 +3,5 @@ package com.example.learnit.ui.feature.courses.quiz.model
 data class QuizResultModel(
     val success: Boolean,
     val message: String,
-    val score: Int,
+    val score: Float,
 )
